requirements.txt
psutil
requests
pyperclip
Pillow
